feat: add DeBERTa-v3 and ModernBERT transformers for text classification without tests. Refactor transformers
