+++ DashAI/back/initial_components.py
@@ -124,10 +124,12 @@ from DashAI.back.metrics.translation.chrf import Chrf
 from DashAI.back.metrics.translation.ter import Ter
 
 # Models
+from DashAI.back.models.hugging_face.deberta_v3_transformer import DebertaV3Transformer
 from DashAI.back.models.hugging_face.distilbert_transformer import DistilBertTransformer
 from DashAI.back.models.hugging_face.llama_model import LlamaModel
 from DashAI.back.models.hugging_face.mistral_model import MistralModel
 from DashAI.back.models.hugging_face.mixtral_model import MixtralModel
+from DashAI.back.models.hugging_face.modernbert_transformer import ModernBertTransformer
 from DashAI.back.models.hugging_face.opus_mt_en_es_transformer import (
     OpusMtEnESTransformer,
 )
@@ -262,6 +264,8 @@ def get_initial_components():
         RandomForestClassifier,
         RandomForestRegression,
         DistilBertTransformer,
+        ModernBertTransformer,
+        DebertaV3Transformer,
         OpusMtEnESTransformer,
         BagOfWordsTextClassificationModel,
         RidgeRegression,
