+++ DashAI/back/models/hugging_face/distilbert_transformer.py
@@ -1,9 +1,5 @@
 """DashAI implementation of DistilBERT model for english classification."""
 
-from typing import TYPE_CHECKING, Any, Union
-
-from sklearn.exceptions import NotFittedError
-
 from DashAI.back.core.schema_fields import (
     BaseSchema,
     enum_field,
@@ -13,14 +9,10 @@ from DashAI.back.core.schema_fields import (
     schema_field,
 )
 from DashAI.back.core.utils import MultilingualString
-from DashAI.back.models.text_classification_model import TextClassificationModel
+from DashAI.back.models.hugging_face.base_text_classification_transformer import (
+    HuggingFaceTextClassificationTransformer,
+)
 from DashAI.back.models.utils import GPU_OR_CPU, GPU_OR_CPU_PLACEHOLDER
-from DashAI.back.types.categorical import Categorical
-
-if TYPE_CHECKING:
-    from pathlib import Path
-
-    from DashAI.back.dataloaders.classes.dashai_dataset import DashAIDataset
 
 
 class DistilBertTransformerSchema(BaseSchema):
@@ -166,7 +158,7 @@ class DistilBertTransformerSchema(BaseSchema):
     )  # type: ignore
 
 
-class DistilBertTransformer(TextClassificationModel):
+class DistilBertTransformer(HuggingFaceTextClassificationTransformer):
     """Pre-trained transformer DistilBERT allowing English text classification.
 
     DistilBERT is a small, fast, cheap and light Transformer model trained by
@@ -191,308 +183,5 @@ class DistilBertTransformer(TextClassificationModel):
     COLOR: str = "#96008E"
     ICON: str = "Psychology"
     SCHEMA = DistilBertTransformerSchema
-
-    def __init__(self, model=None, **kwargs):
-        """Initialize the transformer model.
-
-        The process includes the instantiation of the pre-trained model and the
-        associated tokenizer.
-        """
-
-        self.num_labels = kwargs.pop("num_labels", None)
-
-        kwargs = self.validate_and_transform(kwargs)
-
-        from transformers import AutoTokenizer
-
-        self.model_name = "distilbert-base-uncased"
-
-        self.tokenizer = AutoTokenizer.from_pretrained(self.model_name)
-
-        self.log_train_every_n_epochs = kwargs.get("log_train_every_n_epochs", 1)
-        self.log_train_every_n_steps = kwargs.get("log_train_every_n_steps", None)
-        self.log_validation_every_n_epochs = kwargs.get(
-            "log_validation_every_n_epochs", 1
-        )
-        self.log_validation_every_n_steps = kwargs.get(
-            "log_validation_every_n_steps", None
-        )
-
-        self.training_args_params = {
-            "num_train_epochs": kwargs.get("num_train_epochs", 2),
-            "learning_rate": kwargs.get("learning_rate", 5e-5),
-            "weight_decay": kwargs.get("weight_decay", 0.01),
-        }
-        self.batch_size = kwargs.get("batch_size", 16)
-        self.device = kwargs.get("device")
-
-        if model is not None:
-            self.model = model
-            if self.num_labels is not None and hasattr(self.model, "config"):
-                self.model.config.num_labels = self.num_labels
-                if self.num_labels > 1:
-                    self.model.config.problem_type = "single_label_classification"
-        else:
-            from transformers import AutoConfig, AutoModelForSequenceClassification
-
-            model_config = AutoConfig.from_pretrained(self.model_name)
-            if self.num_labels is not None:
-                model_config.num_labels = self.num_labels
-                if self.num_labels > 1:
-                    model_config.problem_type = "single_label_classification"
-            # Fallback: num_labels will be determined in fit().
-            self.model = AutoModelForSequenceClassification.from_pretrained(
-                self.model_name, config=model_config
-            )
-
-        self.fitted = False
-        self.encodings = {}  # Store encodings for categorical columns
-
-    def train(self, x_train, y_train, x_validation, y_validation):
-        import shutil
-
-        import torch
-        from transformers import (
-            AutoConfig,
-            AutoModelForSequenceClassification,
-            DataCollatorWithPadding,
-            Trainer,
-            TrainingArguments,
-        )
-
-        from DashAI.back.models.hugging_face.metrics_callback import MetricsCallback
-
-        output_column_name = y_train.column_names[0]
-
-        if self.num_labels is None:
-            self.num_labels = len(y_train.unique(output_column_name))
-            config = AutoConfig.from_pretrained(
-                self.model_name, num_labels=self.num_labels
-            )
-            if self.num_labels > 1:
-                config.problem_type = "single_label_classification"
-            self.model = AutoModelForSequenceClassification.from_pretrained(
-                self.model_name, config=config
-            )
-
-        # Train dataset preparation
-        x_train_prepared = self.prepare_dataset(x_train, is_fit=True)
-        y_train_prepared = self.prepare_dataset(y_train, is_fit=True)
-        train_dataset = x_train_prepared.add_column(
-            "label", y_train_prepared[output_column_name]
-        )
-
-        # Validation dataset preparation
-        x_validation_prepared = self.prepare_dataset(x_validation)
-        y_validation_prepared = self.prepare_dataset(y_validation)
-        validation_dataset = x_validation_prepared.add_column(
-            "label", y_validation_prepared[output_column_name]
-        )
-
-        # Get number of epochs from training args
-        num_epochs = self.training_args_params.get("num_train_epochs", 2)
-
-        can_use_fp16 = torch.cuda.is_available() and self.device.lower() == "gpu"
-        training_args_obj = TrainingArguments(
-            output_dir="DashAI/back/user_models/temp_checkpoints_distilbert",
-            save_strategy="epoch",
-            per_device_train_batch_size=self.batch_size,
-            per_device_eval_batch_size=self.batch_size,
-            eval_strategy="no",
-            use_cpu=self.device.lower() != "gpu",
-            fp16=can_use_fp16,
-            **self.training_args_params,
-        )
-
-        data_collator = DataCollatorWithPadding(tokenizer=self.tokenizer)
-
-        # Initialize the custom callback with epoch information
-        metrics_callback = MetricsCallback(
-            model_instance=self,
-            x_train=x_train,
-            y_train=y_train,
-            x_val=x_validation,
-            y_val=y_validation,
-            total_epochs=num_epochs,
-            log_training_every_n_epochs=self.log_train_every_n_epochs,
-            log_training_every_n_steps=self.log_train_every_n_steps,
-            log_val_every_n_epochs=self.log_validation_every_n_epochs,
-            log_val_every_n_steps=self.log_validation_every_n_steps,
-        )
-
-        trainer = Trainer(
-            model=self.model,
-            args=training_args_obj,
-            train_dataset=train_dataset,
-            eval_dataset=validation_dataset,
-            data_collator=data_collator,
-            callbacks=[metrics_callback],
-        )
-
-        self.fitted = True
-        trainer.train()
-
-        shutil.rmtree(
-            "DashAI/back/user_models/temp_checkpoints_distilbert", ignore_errors=True
-        )
-        return self
-
-    def predict(self, x_pred: "DashAIDataset"):
-        """Predict with the fine-tuned model.
-
-        Parameters
-        ----------
-        x_pred : DashAIDataset
-            Dataset with text data.
-
-        Returns
-        -------
-        List
-            List of predicted probabilities for each class.
-        """
-
-        if not self.fitted:
-            raise NotFittedError(
-                f"This {self.__class__.__name__} instance is not fitted yet. Call 'fit'"
-                " with appropriate arguments before using this estimator."
-            )
-
-        pred_dataset = self.prepare_dataset(x_pred)
-
-        import numpy as np
-        from torch.utils.data import DataLoader
-        from transformers import DataCollatorWithPadding
-
-        data_collator = DataCollatorWithPadding(tokenizer=self.tokenizer)
-        text_columns = [col for col in x_pred.column_names if col != "label"]
-        if len(text_columns) != 1:
-            raise ValueError(f"Expected exactly one text column, found: {text_columns}")
-
-        pred_loader = DataLoader(
-            pred_dataset.remove_columns(text_columns[0]),
-            batch_size=self.batch_size,
-            collate_fn=data_collator,
-        )
-
-        probabilities = []
-
-        for batch in pred_loader:
-            inputs = {
-                k: v.to(self.model.device) for k, v in batch.items() if k != "labels"
-            }
-
-            outputs = self.model(**inputs)
-            probs = outputs.logits.softmax(dim=-1)
-            probabilities.append(probs.detach().cpu().numpy())
-
-        return np.vstack(probabilities)
-
-    def prepare_dataset(
-        self, dataset: "DashAIDataset", is_fit: bool = False
-    ) -> "DashAIDataset":
-        """Apply the model transformations to the dataset.
-
-        Parameters
-        ----------
-        dataset : DashAIDataset
-            The dataset to be transformed.
-        is_fit : bool
-            Whether this is for fitting (True) or prediction (False).
-
-        Returns
-        -------
-        DashAIDataset
-            The prepared dataset ready to be converted to
-            an accepted format in the model.
-        """
-        from DashAI.back.dataloaders.classes.dashai_dataset_utils import (
-            apply_categorical_label_encoder,
-            categorical_label_encoder,
-        )
-
-        has_categorical = any(
-            isinstance(col_type, Categorical) for col_type in dataset.types.values()
-        )
-
-        if has_categorical:
-            if is_fit:
-                dataset, encodings = categorical_label_encoder(dataset)
-                self.encodings.update(encodings)
-            else:
-                dataset = apply_categorical_label_encoder(dataset, self.encodings)
-            return dataset
-        else:
-            return self.tokenize_data(dataset)
-
-    def tokenize_data(self, dataset: "DashAIDataset") -> "DashAIDataset":
-        """Tokenize the input data.
-
-        Parameters
-        ----------
-        dataset : DashAIDataset
-            Dataset with the input data to preprocess.
-
-        Returns
-        -------
-        DashAIDataset
-            Dataset with the tokenized input data.
-        """
-        text_columns = [
-            col
-            for col in dataset.column_names
-            if not isinstance(dataset.types.get(col), Categorical)
-        ]
-        if len(text_columns) != 1:
-            raise ValueError(f"Expected exactly one text column, found: {text_columns}")
-
-        return dataset.map(
-            lambda batch: self.tokenizer(
-                batch[text_columns[0]], truncation=True, padding=True, max_length=512
-            ),
-            batched=True,
-        )
-
-    def save(self, filename: Union[str, "Path"]) -> None:
-        from transformers import AutoConfig
-
-        self.model.save_pretrained(filename)
-        config = AutoConfig.from_pretrained(filename)
-        config.custom_params = {
-            "num_train_epochs": self.training_args_params.get("num_train_epochs"),
-            "batch_size": self.batch_size,
-            "learning_rate": self.training_args_params.get("learning_rate"),
-            "device": self.device,
-            "weight_decay": self.training_args_params.get("weight_decay"),
-            "num_labels": self.num_labels,
-            "fitted": self.fitted,
-        }
-
-        config.save_pretrained(filename)
-
-    @classmethod
-    def load(cls, filename: Union[str, "Path"]) -> Any:
-        from transformers import AutoConfig, AutoModelForSequenceClassification
-
-        config = AutoConfig.from_pretrained(filename)
-        custom_params = getattr(config, "custom_params", {})
-
-        model = AutoModelForSequenceClassification.from_pretrained(
-            filename, num_labels=custom_params.get("num_labels")
-        )
-        loaded_model = cls(
-            model=model,
-            model_name=config.model_type,
-            num_labels=custom_params.get("num_labels"),
-            num_train_epochs=custom_params.get("num_train_epochs"),
-            batch_size=custom_params.get("batch_size"),
-            learning_rate=custom_params.get("learning_rate"),
-            device=custom_params.get("device"),
-            weight_decay=custom_params.get("weight_decay"),
-            log_train_every_n_epochs=None,
-            log_train_every_n_steps=None,
-            log_validation_every_n_epochs=None,
-            log_validation_every_n_steps=None,
-        )
-        loaded_model.fitted = custom_params.get("fitted")
-
-        return loaded_model
+    MODEL_NAME: str = "distilbert-base-uncased"
+    TEMP_CHECKPOINT_DIR: str = "DashAI/back/user_models/temp_checkpoints_distilbert"
